+++ src/apps/peer/udpserver.h
@@ -56,7 +56,7 @@ struct server_info {
 
 //////////////////////////////
 
-server_type *start_udp_server(int verbose, const char *ifname, char **local_addresses, size_t las, int port);
+server_type *start_udp_server(int verbose, const char *ifname, char **local_addresses, size_t las, uint16_t port);
 
 void run_udp_server(server_type *server);
